Halo Reach and Halo 3 Merged into Gen3 Cache Formats

+++ Tags/MantleLib/CacheInterface/CacheVersions/halo3odst/halo3odst_cache_file.h
@@ -11,7 +11,7 @@ public:
 	virtual bool save_map() final;
 	virtual bool is_loading() const final;
 	virtual uint64_t get_base_virtual_address() const final;
-	virtual uint64_t convert_page_offset(uint64_t page_offset) const final;
+	virtual uint64_t convert_page_offset(uint32_t page_offset) const final;
 	virtual uint32_t get_tag_count() const final;
 	virtual uint32_t get_tag_group_count() const final;
 	virtual uint32_t get_string_id_count() const final;
